Lead details are shown using LeadDetailsContainerViewController.
Why?
The previous approach was to use the LeadDetailsPagingViewController to display the lead details once the lead is selected in the LeadsTableViewController. The problem with this approach was that the parent navigation controller was not being set and hence could not implement any customisations.

Now.
Lead details are now shown using the LeadDetailsContainerViewController. LeadDetailsPagingViewController is now added as a child view controller to LeadDetailsContainerViewController. The view of the LeadDetailsPagingViewController is embedded into the LeadDetailsContainerViewController’s view. A lead’s tasks, opportunities, contacts, messages and status are presented to the user. The user can edit the lead using the LeadDetailsContainerViewController. LeadsTableViewController is now given as the change delegate to AddLeadViewController and the LeadDetailsContainerViewController. So once a addition/change happens to a lead, the LeadsTableViewController gets the message to reload the data in the table view controller.

Minor Fixes.
* TableViews of LeadTasksTableViewController, LeadOpportunitiesTableViewController, LeadContactsTableViewController, LeadMessagesTableViewController and LeadStatusTableViewController are styled.
* NavigationController passed between paging view controllers is refactored out.