Feature to choose the user to whom the task is to be assigned.
What?
When adding a new task from the LeadDetailsContainerViewController, the user need to choose to whom the task is to be assigned from a list of users. ChooseUserTableViewController displays the list of users where you need to select to whom the task should be assigned.

How?
The ChooseUserTableViewController sends a message to a delegate of type UserChoosenDelegate when an user is selected. The delegate then displays the name of the selected user.