Adding tasks to a lead.
Tasks can be added to a lead from the LeadDetailsContainerViewController. The various properties that can be set of a task include task description, assigned to user and expiry date. The task can be in the following two states, closed and opened. Once the task is done, the user can close the task and they can re-open when desired. The task can also be deleted.

A new task can be added from the LeadDetailsContainerViewController. The user needs to input task description and expiry date. The assigned to user can be chosen with the help of the ChooseUserViewController. Once the task is added, a message is sent to its delegate. The LeadTasksTableViewController is the delegate and once it receives the message, it refreshes itself.

* Other minor refactoring