Contacts can be added to a lead.
The functionality to add contacts is enabled. The contact’s name, email and phone number can be entered using the AddLeadContactViewController. The contacts are displayed using the LeadContactsTableViewController. The contacts can be deleted using the same view controller. When a specific cell is clicked, various communication options are presented to the user. The user can call, text message or email the contact from within the app. A delegate is implemented by the LeadContactsTableViewController so as to refresh itself when a new contact is added.

* Contact data model encapsulates a lead contact
* Various assets are added to present various communication options