Reactive refresh when an user is added using the AddUserViewController.
What?
When an user is added using the AddUserViewController, the current approach is to use the delegate protocol to communicate to the UsersTableViewController to reload the table view. This is not a pure “reactive” solution. Hence this change does not refresh the UsersTableViewController through the delegate call, but listens for the new/edit notifications of the Users ResultSet (Realm data structure to hold data) and when the notifications get triggered, the table view is refreshed.

Why?
The reactive solution looks more clean than the delegate way of communicating. The delegate is still called to pass other information that is needed but the reloading of table view is not done here.