Reactive way to refresh tasks table view when a task is deleted/closed/re-opened
What?
When a task is deleted/closed/re-opened, the table view is not refreshed. The corresponding result sets are listened for notifications and the table view is refreshed in the notification block.

Why?
The refresh is only done by listening to the notifications of the respective result sets. This is clearer and more reactive approach.

Minor Fixes:
* isScrollEnabled property of the PagingViewController is set to false. This is so since we allow the user to delete or close/open a task and the scrolling can get in the way of UX