Create a new Lead Opportunity / display Lead Opportunity
What?
Create a new lead opportunity from the LeadDetailsContainerViewController. LeadOpportunitiesTableViewController is used to display the lead opportunity created. When creating a new opportunity, the user needs to enter details like description, value, confidence percentage, assigned to user etc. Displaying a opportunity shows details like value, assigned to user, percentage and status.

LeadOpportunityTableViewCell is used to display lead opportunities details.

Opportunity is the model used to define a opportunity and its various properties.

AddLeadOpportunityDelegate’s didFinishAddingLeadOpportunity method is called when a new lead opportunity is added.

Notifications are listened to when a new lead opportunity is added so as to add a new opportunity to LeadOpportunitiesTableViewController and OpportunitiesTableViewController.

Why?
LeadOpportunityTableViewCell is used since custom table view cell can’t display all the information that we need.

AddLeadOpportunityDelegate’s method is called to present a window to the user to perform custom actions.

Insertions, deletions and modifications notifications are listened to display the changes.

minor fixes.
* sorted(byProperty:) is changed to sorted(byKeyPath:) as the previous method is deprecated