Create a new Opportunity / display Opportunity
What?
Create a new lead opportunity from the OpportunitiesTableViewController. It is also used to display the lead opportunity created. When creating a new opportunity, the user needs to enter details like description, lead, value, confidence percentage, assigned to user etc. Displaying a opportunity shows details like value, assigned to user, percentage and status.

OpportunityTableViewCell is used to display lead opportunities details. The lead short name is also displayed in the cell.

Opportunity is the model used to define a opportunity and its various properties.

AddOpportunityDelegate’s didFinishAddingOpportunity method is called when a new lead opportunity is added.

Notifications are listened to when a new lead opportunity is added so as to add a new opportunity to LeadOpportunitiesTableViewController and OpportunitiesTableViewController.

Why?
OpportunityTableViewCell is used since custom table view cell can’t display all the information that we need.

AddOpportunityDelegate’s method is called to present a window to the user to perform custom actions.

Insertions, deletions and modifications notifications are listened to display the changes.