Use various image assets as is without re-sizing
Images used for the various UITabBarItems and others in the app are to be used as is without re-sizing.

Images used for UITabBarItems, LeadContact options, LeadMessages should use the new images as is.

Why?
Earlier, when we re-sized the various images to be used, the image becomes distorted in the device/simulator. Hence the decision was taken to create the image with the needed dimensions using a graphic program (We use Sketch) and then use them as is in the app.

Minor fixes:
- func styleTabBarItem(tabBarItem:, imageName:) is refactored to func styleTabBarItem(tabBarItem:, imageName:, to:). The user can pass in a custom CGSize to the to parameter. By default, to’s value is CGSize(width: 30.0, height: 30.0)
- Various Strings, Color Codes and Font names in ProjectCloseStrings, ProjectCloseColors and ProjectCloseFonts