Use custom made MetricTableViewCell to display various lead reports
MetricTableViewCell to be used to display various reports of a lead. The ReportTableViewCell will display the metric as a percentage graph on the left and WoW (Week on Week) as a bar chart on the right. Each section will have the lead name as the header and the various metrics below it.

Why?
The design of MetricTableViewCell brings out more details than the previously used ReportTableViewCell. The information is clear and uses whitespace much better.

Minor Fixes:
- LeadId is added to the Contact model. Each contact should be associated with a lead
- Tasks, Opportunities, Contacts in LeadTasksTableViewController, LeadOpportunitiesTableViewController, ContactsTableViewController respectively filters out entries by leadId. A NSPredicate condition is passed to filter out the non lead entries
- When a PagingViewController is used, the lazily loading property is specified as one instead of all. This is because when lazy loading all, various UITableView data source methods are called incorrectly. This might be a problem with the third party PagingViewController library
- Various font and color entries