Use dequeueReusableCell methods to get a new UITableViewCell
Initially, a new UITableViewCell (custom or built-in) was got using the init method of the respective UITableViewCell. This is not efficient and moves away from the  best practices followed in iOS development. All UITableViewCell needs to be got using either the dequeueReusableCell(withIdentifier:) or dequeueReusableCell(withIdentifier:, for:) method. Custom cells can be got using the latter method and built-in cells with non default styles can be got using the former method.

Minor fixes:
- Various comments have been removed