Move creation of percentageLayer to ReportTableViewCell
Earlier, the percentageLayer denoting the percentage change was created and initialised when the cell was asked by the UITableViewDelegate. Since, this was not tied to a particular UITableViewCell, pervious values were not being overwritten and inconsistent values were being shown.

To prevent this from happening, the layer is now being created in the ReportTableViewCell and hence is now tied to the UITableViewCell and each cell has its own value.