Welcome screen need to be presented to the user
WelcomeViewController is introduced to the user asking for his action if they need to register (new user) or login (existing user). The logo along with the company name is shown to the user in the same view controller.

LoginViewController is changed to be consistent with WelcomeViewController in terms of theming and style. The logo along with the company names is also shown in this view controller.

The modified flow now is WelcomeViewController->LoginViewController->MainTabBarViewController.

Minor Fixes:
- Added images for the MainLogo
- Changed the background image for LoginViewController
- Various color, font and string entries