Change login screen design for ProjectClose
New login screen is developed for the user to log in. The old designs were a bit cluttered. The new design is simple and clean. Initially a welcome screen was presented to the user and the user had to choose between logging and registering. Now only the login screen is presented to the user.

Why?
The decision is taken to give the user a clear picture and not take them around lots of screens.

Minor Fixes:
- Color, Fonts and Strings entries