Super user should always be present
Since this is app is for demo purposes, some short cuts were taken to create models. Some models had the assumption that an user was already present. To support these workflows, a user is created during the first run of the application. When the user logs in initially, there is a check if this super user is present. If yes, nothing is done. If not, the super user is created.